pyproject.toml
[project]
name = "your-first-financial-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dotenv>=0.9.9",
    "openai>=2.16.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "sniffio>=1.3.1",
    "zai-sdk>=0.2.1",
]
